Rename and move notices index
